Minor adjustments to user settings
